feat: useProfileApi 기능 분리
