+++ src/pages/MyCareer/ProfileEditPage/ProfileEditPage.jsx
@@ -3,33 +3,17 @@ import UnderlineButton from '../../../components/common/Button/UnderlineButton/U
 import ProfileSetting from '../../../components/common/ProfileSetting/ProfileSetting';
 import * as S from './styled/styled';
 import AccountPopUp from '../../../components/common/Popups/AccountPopUp/AccountPopUp';
-import { profileInitialData } from '../../../data/profileData';
 import { useProfile } from '../../../hooks/useProfile';
-import { useMutation, useQuery } from '@tanstack/react-query';
-import { getProfile, modifyProfile } from '../../../apis/profileApi';
 import { useAuthStore } from '../../../store/authStore';
+import { useEditProfile, useFetchProfile } from '../../../apis/profile/useProfileApi';
 
 const ProfileEditPage = () => {
     const [isPopUp, setIsPopUp] = useState(false);
-    const { fetchUser } = useAuthStore();
-    const { data, isLoading, isError, error } = useQuery({
-        queryKey: ['profile'],
-        queryFn: getProfile,
-        onSuccess: (data) => {
-            fetchUser(data);
-        },
-    });
-    const { canSave, emailError, profile, handleProfileFieldChange } = useProfile(profileInitialData);
+    const { user } = useAuthStore();
+    const { isLoading, isError, error } = useFetchProfile();
+    const { canSave, emailError, profile, handleProfileFieldChange } = useProfile(user);
 
-    const mutation = useMutation({
-        mutationFn: modifyProfile,
-        onSuccess: () => {
-            fetchUser(profile);
-        },
-        onError: (error) => {
-            console.error(error);
-        },
-    });
+    const mutation = useEditProfile(profile);
 
     const handlePopUpOpen = () => {
         setIsPopUp(true);
@@ -38,14 +22,6 @@ const ProfileEditPage = () => {
     const handlePopUpClose = () => {
         setIsPopUp(false);
     };
-
-    if (isLoading) {
-        return <div>Loading...</div>;
-    }
-
-    if (isError) {
-        return <div>Error: {error.message}</div>;
-    }
 
     const handleSave = () => {
         if (!canSave) {
@@ -60,11 +36,19 @@ const ProfileEditPage = () => {
         mutation.mutate(profile);
     };
 
+    if (isLoading) {
+        return <div>Loading...</div>;
+    }
+
+    if (isError) {
+        return <div>Error: {error.message}</div>;
+    }
+
     return (
         <S.EditContainer>
             <S.NoticeWrapper>
-                <S.NoticeTitle>{profileInitialData.name} 님의 프로필</S.NoticeTitle>
-                <S.NoticeDetail>{profileInitialData.name} 님의 정보를 수정해주세요!</S.NoticeDetail>
+                <S.NoticeTitle>{user.name} 님의 프로필</S.NoticeTitle>
+                <S.NoticeDetail>{user.name} 님의 정보를 수정해주세요!</S.NoticeDetail>
             </S.NoticeWrapper>
             <S.ContentWrapper>
                 <ProfileSetting
